requirements.txt
requests>=2.31.0
discord.py>=2.3.0
python-telegram-bot>=20.0
cloudscraper>=1.2.71
DrissionPage>=4.0.0
schedule>=1.2.0

